Create Button.js
This is the JavaScript that makes sections of the screen clickable. The Button constructor takes four ints, bounding the rectangle, and a function pointer, which is the action taken when the button is clicked. Buttons can be toggled on and off with the setEnabled(Boolean) method. When a position on the screen is clicked, the appropriate Button will react.