Update CoffeeShop.html
Alright. Did some more work, so we can now swap scenes. Buttons for the inventory/phone will need to be added to the persistent buttons list, so that they will be available in every scene. I envision creating a function that draws the HUD (inventory, phone, etc.) on top of image currently being displayed, so that can also be persistent across multiple scenes.