Updated the game with the capibility to save the users progress.
Also added an icon to go on the browser tab. I chose the C in the CyberEDU logo.

Note the server URL is hardcoded into the game in four places as jrhansf.student.umd.edu. If you are attempting to run from a different machine,
you'll need to change that.