The beginnings of the Second Iteration of the CyberEdu game...
Removed all of the old code from the repository.

Added in the beginnings of the new code (because I was afraid of losing it if my laptop's hard drive went up; it ran Chkdisk upon startup a few days ago). So far, I have added in the internet browser, the dialog box, and the phone, although it has no apps. I need to add the computer filesystem, the email system, and some type of map before I open the floodgates to code up all the modules.

CAS authentication and the save/load feature will also need to be added, but that can wait because it is annoying to have to repeatedly delete your own save file from the server. It should be fairly easy to add at any time, since I have been designing with it in mind. (Just read in the game object from a JSON file instead of hardcoding it, and write it to a JSON file upon disconnection).

Not all of the resources (images, videos, audio files) have been added yet.