Added background music to the game.
To add music to a scene, you must add the ID of the audio file to the background_music map maintained within the game object.

This should be done in the load method of a scene, like this:
game.background_music["coffee_shop"] = "audio/coffee_shop_background_music";

Note that if you would like this music to play on multiple screens, you must do this for ALL of those screens.