Added scroll buttons to the email application.
Added a function which changes the volume on the client side -- currently no GUI component to access that.

Added verification code to the beginning of socket.on('click', function (button) { ... }), to (hopefully) prevent someone from sending a malicious event via the console -- this verification is supposed to ignore any click event that occurs to an element that is not on screen.