Modified the Cafe image in Paint. (The people's laptop's got removed.)
Coffee Shop now uses the Todo list app.

When drawing a dialog over a screen with buttons/textFields, the text in those buttons/textFields is now added to the underlying screen. (Previously, that did not occur).

Fixed bug where the leaderboard generation code would throw an exception if it read a save file which had not yet loaded the scenes.