Added in the new map image, but did not add in the individual map icons.
Removed voice from the coffee shop. TODO: remove completely from game.

Fixed final module image to the latest updated version (hitboxes were not affected by this change).

Changed how task #2 was implemented (which was assigned to the freshmen). TODO: fix messages related to unlocking locations.

Emails are now marked as read when they are backed out of (or deleted), not the instant they are actually read.